+++ src/integration-test/java/com/commercetools/sync/integration/externalsource/categories/updateactionutils/ChangeParentIT.java
@@ -10,6 +10,7 @@ import io.sphere.sdk.categories.commands.CategoryCreateCommand;
 import io.sphere.sdk.categories.commands.updateactions.ChangeParent;
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.models.LocalizedString;
+import io.sphere.sdk.models.ResourceIdentifier;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -106,7 +107,7 @@ class ChangeParentIT {
         final CategoryDraft newCategory = CategoryDraftBuilder
             .of(LocalizedString.of(Locale.ENGLISH, "classic furniture"),
                 LocalizedString.of(Locale.ENGLISH, "classic-furniture"))
-            .parent(oldCategory.getParent().toResourceIdentifier())
+            .parent(ResourceIdentifier.ofId(oldCategory.getParent().getId()))
             .build();
 
         // Build change parent update action