+++ docs/RELEASE_NOTES.md
@@ -36,8 +36,16 @@
 [Javadoc](https://commercetools.github.io/commercetools-sync-java/v/1.2.0/) | 
 [Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/1.2.0)
 
-- 🚧 **Breaking Changes** (1)
+- 🚧 **Breaking Changes** (2)
     - **ProductType Sync** - Removed the unneeded `AttributeDefinitionCustomBuilder` which was an exposed but internal helper. [#377](https://github.com/commercetools/commercetools-sync-java/issues/377). 
+    - **Commons** - `SyncUtils#replaceReferenceIdWithKey` now renamed to `SyncUtils#getReferenceWithKeyReplaced`. [#349](https://github.com/commercetools/commercetools-sync-java/issues/349)
+
+- 🎉 **New Features** (2)
+    - **Commons** - Added the new `CommonTypeUpdateActionUtils#buildUpdateActionForReferences` which is used for comapring references/resourceIdentifiers and buiding an update action if needed. [#349](https://github.com/commercetools/commercetools-sync-java/issues/349)
+    - **Commons** - Added the new `SyncUtils#getResourceIdentifierWithKeyReplaced` util. [#349](https://github.com/commercetools/commercetools-sync-java/issues/349)
+    
+- 🐞 **Bug Fixes** (1)
+    - **Commons** - Fixed a bug where references and resource identifiers were not being compared correctly. [#349](https://github.com/commercetools/commercetools-sync-java/issues/349)
 
 - 🛠️ **Enhancements** (12)
     - **Commons** - Bumped commercetools-jvm-sdk to version [1.42.0](http://commercetools.github.io/commercetools-jvm-sdk/apidocs/io/sphere/sdk/meta/ReleaseNotes.html#v1_42_0).